% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{load_key}
\alias{load_key}
\title{Load an API key}
\usage{
load_key(keyname = "key")
}
\arguments{
\item{keyname}{An optional string containing the name of the key}
}
\value{
A string containing the loaded key.
}
\description{
Load an API key string from a config file created by \code{save_key()}.
}
\examples{
key <- load_key("openai")

}
