% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt.R
\name{prompt_openai}
\alias{prompt_openai}
\title{Prompt OpenAI API}
\usage{
prompt_openai(prompt, model, context, key, temperature = 0)
}
\arguments{
\item{prompt}{A required string containing the prompt to give the LLM.}

\item{model}{A required string containing the model name in OpenAI's API.}

\item{context}{Either NULL or a string containing the system context to give
the LLM (default = NULL).}

\item{key}{A required string containing your API key. See \code{save_key()} and
\code{load_key()} for related convenience functions.}

\item{temperature}{A number representing how random vs. deterministic the
output is. Set to 0 to always get the same response.}
}
\value{
A string containing the LLM's response to your prompt.
}
\description{
Prompt local large language model from OpenAI using API.
}
\examples{
prompt_openai(
  prompt = "Introduce yourself.",
  model = "gpt-4o-mini",
  context = "Always answer in rhymes.",
  key = load_key()
)
}
