% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{remove_key}
\alias{remove_key}
\title{Remove a saved API key}
\usage{
remove_key(keyname)
}
\arguments{
\item{keyname}{A string containing the name of the key to remove.}
}
\description{
Delete from your computer the API key stored under \code{name}. Find all stored
API keys using \code{find_keys()}.
}
\examples{
remove_key("openai")

}
