% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{save_key}
\alias{save_key}
\title{Save an API key}
\usage{
save_key(key, keyname = "key")
}
\arguments{
\item{key}{A string containing the API key}

\item{keyname}{An optional string containing the name of the key (to store
multiple keys for use with \code{load_key()}).}
}
\description{
Save an API key string to a config file for easier use later.
}
\examples{
save_key("MY-KEY-HERE", keyname = "openai")
}
